+++ doc/whatsnew.rst
@@ -10,6 +10,11 @@ A catalog of new features, improvements, and bug-fixes in each release.
 v0.8.dev
 --------
 
+- Enhance :func:`pyriemann.utils.mean.mean_ale` adding ``init`` parameter,
+  and add function ``check_init()`` useful to all ajd and mean functions allowing initialization. :pr:`328` by :user:`qbarthelemy`
+
+- Enhance :func:`pyriemann.utils.mean.mean_covariance` to support "power" and "poweuclid" metrics. :pr:`329` by :user:`qbarthelemy`
+
 - Add an example using fNIRS data with a new estimator called `BlockKernels` for classifying HbO and HbR signals. :pr:`323` by :user:`timnaher`
 
 v0.7 (October 2024)